tracking down bugs

+++ libraries/AP_IRLock/AP_IRLock_SITL.h
@@ -28,7 +28,7 @@ private:
      */
     struct irlock_packet {
             uint64_t timestamp;
-            size_t num_targets;
+            uint16_t num_targets;
             float pos_x;
             float pos_y;
             float size_x;
